+++ changelog.md
@@ -4,6 +4,8 @@
 
 ### Version 5.0.0
 
+* `api-scanner`: Compatibility fix for Gradle 5.x.
+
 * `cordformation`: Unload `SecurityProvider` instances installed by Network Bootstrapper to prevent linkage exceptions later.
 
 * `cordformation`: Revert `deployNodes` back to using the runtime classpath again. Anything needed solely by `deployNodes` can be added to Gradle's `runtimeOnly` configuration instead.