+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## Version 4
 
+### Version 4.0.44
+
 ### Version 4.0.43
 
 * `cordformation`: Add support for network parameter overrides.